+++ index.html
@@ -239,9 +239,7 @@
 <script src="leaflet/leaflet.js"></script>
 <script type="module" src="js/map.js"></script>
 <script type="module" src="js/main.js"></script>
-<script type="module" src="js/data.js"></script>
 <script type="module" src="js/activation.js"></script>
 <script type="module" src="js/validity.js"></script>
-<script type="module" src="js/api.js"></script>
 </body>
 </html>
