+++ js/main.js
@@ -1,99 +1,6 @@
 import {
   SIMILAR_ADVERT_COUNT,
-  MAX_PRICE,
-  MAX_ROOMS,
-  MAX_GUESTS,
-  LAT_MIN,
-  LAT_MAX,
-  LNG_MIN,
-  LNG_MAX,
-  FEATURES,
-  PHOTOS,
-  CHECKIN_TIMES,
-  CHECKOUT_TIMES,
-  TITLES,
-  TYPES
+  createAdvert
 } from './data.js';
-
-export const typesRussian = {
-  palace: 'дворец',
-  flat: 'квартира',
-  house: 'дом',
-  bungalow: 'бунгало',
-  hotel: 'отель',
-};
-
-function getRandomIntegerInRange(min, max) {
-  if (max <= min) {
-    return 'Второе число не может быть меньше первого!';
-  }
-  if (min < 0 || max < 0) {
-    return 'Нельзя использовать отрицательное число!';
-  }
-  return Math.floor(Math.random() * (max - min + 1)) + min;
-}
-
-
-function getRandomFloat(min, max, num) {
-  if (max <= min) {
-    return 'Второе число не может быть меньше первого!';
-  }
-  if (min < 0 || max < 0) {
-    return 'Нельзя использовать отрицательное число!';
-  }
-  return (Math.random() * (max - min) + min).toFixed(num);
-}
-
-
-function chooseAvatar() {
-  const i = Math.round(Math.random() * (8 - 1) + 1);
-  const link = `img/avatars/user0 + ${i} + .png`;
-  return link;
-}
-
-
-function getFirst(array) {
-  if (!array) {
-    return [];
-  }
-  const n = getRandomIntegerInRange(0, array.length - 1);
-  return array.slice(0, n);
-}
-
-const getRandomArrayElement = (elements) => {
-  return elements[getRandomIntegerInRange(0, elements.length - 1)];
-};
-
-let author;
-let cardElement;
-
-function createAdvert() {
-const tempLat = getRandomFloat(LAT_MIN, LAT_MAX, 5);
-const tempLng = getRandomFloat(LNG_MIN, LNG_MAX, 5);
-  return {
-    author: {
-      avatar: chooseAvatar(),
-    },
-
-    offer: {
-      title: getRandomArrayElement(TITLES),
-      address: `${tempLat} , ${tempLng}`,
-      price: getRandomIntegerInRange(0, MAX_PRICE),
-      type: getRandomArrayElement(TYPES),
-      rooms: getRandomIntegerInRange(0, MAX_ROOMS),
-      guests: getRandomIntegerInRange(0, MAX_GUESTS),
-      checkin: getRandomArrayElement(CHECKIN_TIMES),
-      checkout: getRandomArrayElement(CHECKOUT_TIMES),
-      features: getFirst(FEATURES),
-      description: 'Расскажите подробнее о вашем жилье',
-      photos: getFirst(PHOTOS),
-    },
-
-    location: {
-      lat: tempLat,
-      lng: tempLng,
-    },
-  };
-}
 
 export const similarAdverts = new Array(SIMILAR_ADVERT_COUNT).fill(null).map(() => createAdvert());
